% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwi.R
\name{fwi}
\alias{fwi}
\title{Fire Weather Index applied to multigrids}
\usage{
fwi(multigrid, mask = NULL, lonLim = NULL, latLim = NULL, lat = 46,
  return.all = FALSE, init.pars = c(85, 6, 15), parallel = FALSE,
  max.ncores = 16, ncores = NULL)
}
\arguments{
\item{multigrid}{containing Tm (temperature records in deg. Celsius); H (relative humidity records in \%);
r (last 24-h accumulated precipitation in mm); W (wind velocity records in Km/h). See the Warning section.}

\item{mask}{Optional. Binary (0 an 1) Grid.}

\item{lonLim}{Vector of length = 2, with minimum and maximum longitude coordinates, in decimal degrees,}

\item{latLim}{Same as \code{lonLim} argument, but for latitude.}

\item{lat}{Optional. Latitude of the records (in decimal degrees). Default to 46,
applying the default parameters of the original FWI System, developed in Canada. See details.}

\item{return.all}{Logical. Should all components of the FWI system be returned?. 
Default to FALSE, indicating that only FWI is returned.}

\item{init.pars}{A numeric vector of length 3 with the initialization values for the
FFMC, DMC and DC components, in this order. Default values as proposed by van Wagner (1987).}

\item{parallel}{Logical. Should parallel execution be used?}

\item{max.ncores}{Integer. Upper bound for user-defined number of cores.}

\item{ncores}{Integer number of cores used in parallel computation. Self-selected number of
cores is used when \code{ncpus = NULL} (the default), or when \code{maxcores} exceeds the default \code{ncores} value.}
}
\value{
A vector of the same length as the input vectors (minus possible missing observations),
containing the requested components of the FWI system (either all or just FWI). See details.
}
\description{
Implementation of the Canadian Fire Weather Index System for multigrids
}
\section{Warning}{


The variables composing the input multigrid need to have standard names, as defined by the dictionary
 (their names are stored in the \code{multigrid$Variable$varName} component).
These are: \code{"tas"} for temperature, \code{"tp"} for precipitation, \code{"wss"} for windspeed. In the case of relative humidity,
either \code{"hurs"} or \code{"hursmin"} are accepted, the latter in case of FWI calculations according to the \dQuote{proxy} version
described in Bedia \emph{et al} 2014.
}

\section{Daylength adjustment factors}{
 
By default, the function applies the original FWI daylength adjustment factors for DC and DMC (van Wagner 1987),
 although it is possible to adjust them by as a function of latitude through the argument \code{lat}.
The reference values used for each latitudinal range are those indicated in p.71 and Tables A3.1 and A3.2 (p69) in
Lawson and Armitage (2008).
}
\author{
J. Bedia & M. Iturbide, partially based on the original FORTRAN code by van Wagner and Pickett (1985)
}
\references{
Lawson, B.D. & Armitage, O.B., 2008. Weather guide for the Canadian Forest Fire Danger Rating System. Northern Forestry Centre, Edmonton (Canada).

van Wagner, C.E., 1987. Development and structure of the Canadian Forest Fire Weather Index (Forestry Tech. Rep. No. 35). Canadian Forestry Service, Ottawa, Canada.

van Wagner, C.E., Pickett, T.L., 1985. Equations and FORTRAN program for the Canadian forest fire weather index system (Forestry Tech. Rep. No. 33). Canadian Forestry Service, Ottawa, Canada.
}

