% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwiGrid.R
\name{fwiGrid}
\alias{fwiGrid}
\title{Fire Weather Index (FWI) from multigrid}
\usage{
fwiGrid(multigrid, mask = NULL, return.all = FALSE, init.pars = c(85, 6,
  15), parallel = FALSE, max.ncores = 16, ncores = NULL)
}
\arguments{
\item{multigrid}{A multigrid of the variables needed to compute the FWI. See the Warning section}

\item{mask}{Optional. Grid of the land mask to be applied to the data.}

\item{return.all}{Logical. Should all components of the FWI system be returned?. 
Default to FALSE, indicating that only FWI is returned.}

\item{init.pars}{A numeric vector of length 3 with the initialization values for the
FFMC, DMC and DC components, in this order. Default values as proposed by van Wagner (1987).}

\item{parallel}{Logical. Should parallel execution be used?}

\item{max.ncores}{Integer. Upper bound for user-defined number of cores.}

\item{ncores}{Integer number of cores used in parallel computation. Self-selected number of
cores is used when \code{ncpus = NULL} (the default), or when \code{maxcores} exceeds the default \code{ncores} value.}
}
\value{
A grid, containing the requested components of the FWI system (either all or just FWI). See details.
}
\description{
A wrapper of function \code{\link{fwi1D}} to handle multigrids
}
\details{
In order to efficiently handle large datasets, the function internally splits the spatial domain in convenient latitudinal chunks.
}
\section{Daylength adjustment factors}{
 
In this case, daylength adjustment factors are automatically set by the funciton according to the
 reference values indicated in p.71 and Tables A3.1 and A3.2 (p69) in Lawson and Armitage (2008).
}

\section{Warning}{


The variables composing the input multigrid need to have standard names, as defined by the dictionary
 (their names are stored in the \code{multigrid$Variable$varName} component).
These are: \code{"tas"} for temperature, \code{"tp"} for precipitation, \code{"wss"} for windspeed. In the case of relative humidity,
either \code{"hurs"} or \code{"hursmin"} are accepted, the latter in case of FWI calculations according to the \dQuote{proxy} version
described in Bedia \emph{et al} 2014.

Be aware of the required input units.
}
\author{
J. Bedia \& M.Iturbide
}
\references{
Lawson, B.D. & Armitage, O.B., 2008. Weather guide for the Canadian Forest Fire Danger Rating System. Northern Forestry Centre, Edmonton (Canada).

van Wagner, C.E., 1987. Development and structure of the Canadian Forest Fire Weather Index (Forestry Tech. Rep. No. 35). Canadian Forestry Service, Ottawa, Canada.

van Wagner, C.E., Pickett, T.L., 1985. Equations and FORTRAN program for the Canadian forest fire weather index system (Forestry Tech. Rep. No. 33). Canadian Forestry Service, Ottawa, Canada.

Bedia, J., Herrera, S., Camia, A., Moreno, J.M., Gutierrez, J.M., 2014. Forest Fire Danger Projections in the Mediterranean using ENSEMBLES Regional Climate Change Scenarios. Clim Chang 122, 185--199. doi:10.1007/s10584-013-1005-z
}
\seealso{
\code{\link{fwi1D}}
}

