% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwiSkill.R
\name{fwiSkill}
\alias{fwiSkill}
\title{Skill calculation}
\usage{
fwiSkill(obs, fcst, verifun = c(""), prob, threshold = NULL,
  parallel = FALSE, na.rm = TRUE)
}
\arguments{
\item{obs}{Grid of the observation data}

\item{fcst}{Grid of the seasonal forecast data}

\item{verifun}{Name of function to compute verification metric (score, skill score). 
Type help(veriApply) for details.}

\item{prob}{Probability threshold for category forecasts (Type help(veriApply) for details).}

\item{threshold}{Aabsolute threshold for category forecasts (Type help(veriApply) for details).}

\item{parallel}{Logical. Should parallel execution be used?}

\item{na.rm}{Logical. Should incomplete forecasts be used?}
}
\value{
A grid of the skill scores.
}
\description{
Skill calculation with function veriApply
}
\author{
M. Iturbide
}

