% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdas.R
\name{meteoSpain}
\alias{meteoSpain}
\title{Time series of meteo station data from Spain}
\format{A data.frame}
\description{
Daily precipitation, maximum and minimum temperature records.
}
\examples{
data(meteoSpain)
str(meteoSpain)
}
