% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwiSkill.R
\name{plotFwiSkill}
\alias{plotFwiSkill}
\title{Plot grid returned by functionfFwiSkill}
\usage{
plotFwiSkill(grid)
}
\arguments{
\item{grid}{The grid object returned by function fwiSkill}
}
\description{
Plot grid returned by functionfFwiSkill through function plotClimatology from package downscaleR
}
\author{
M.Iturbide
}

