% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wfwibc.R
\name{wfwibc}
\alias{wfwibc}
\title{World Bias Correction}
\usage{
wfwibc(dataset, dictionary = TRUE, members = NULL, season = NULL,
  years = NULL, leadMonth = 1, lat = 46, return.all = FALSE,
  init.pars = c(85, 6, 15), bc.multigrid, method = c("delta", "scaling",
  "eqm", "gqm", "gpqm"), cross.val = c("none", "loocv", "kfold"),
  folds = NULL, window = NULL, scaling.type = c("additive",
  "multiplicative"), wet.threshold = 1, n.quantiles = NULL,
  extrapolation = c("none", "constant"), theta = 0.95, parallel = FALSE,
  max.ncores = 16, ncores = NULL)
}
\arguments{
\item{dataset}{A character string indicating the database to be accessed (partial matching is enabled).
Currently accepted values are "System4_seasonal_15" and "CFSv2_seasonal".}

\item{dictionary}{A logical flag indicating if a dictionary is used for variable homogenization. Default (strongly recommended)
is set to TRUE, meaning that the function will internally perform the necessary homogenization steps to return the standard
variables defined in the vocabulary (e.g. variable transformation, deaccumulation...). See details on data homogenization.}

\item{members}{A vector of integers indicating the members to be loaded. Default to NULL, which loads the default members
(see details for the particularities concerning the CFSv2 dataset). For instance, members=1:5 will retrieve the first five
members of the hindcast. Discontinuous member selection (e.g. members=c(1,5,7)) is allowed. If the requested dataset is not
a forecast or the requested variable is static (e.g. orography) it will be ignored.}

\item{season}{An integer vector specifying the desired season (in months, January = 1 ..., December = 12). Options include
one to several (contiguous) months. For full year selections (not possible for all datasets, e.g. seasonal forecasts),
the argument value must be set to season = 1:12. If the requested variable is static (e.g. orography) it will be ignored.
See details on the definition of temporal slices.}

\item{years}{Optional vector of years to select. Default to all available years. If the requested variable is static (e.g. orography)
it will be ignored. See details on the definition of temporal slices.}

\item{leadMonth}{Integer value indicating the lead forecast time, relative to the first month of season. Note that leadMonth=1 for
season=1 (January) corresponds to the December initialization. Default to 1 (i.e., 1 lead month forecast). If the dataset is not a
forecast or the requested variable is static (e.g. orography) it will be ignored. A message will be printed on screen in the first
case if its value is different from NULL. See details on initialization times.}

\item{lat}{Optional. Latitude of the records (in decimal degrees). Default to 46,
applying the default parameters of the original FWI System, developed in Canada. See details.}

\item{return.all}{Logical. Should all components of the FWI system be returned?.
Default to FALSE, indicating that only FWI is returned.}

\item{init.pars}{A numeric vector of length 3 with the initialization values for the
FFMC, DMC and DC components, in this order. Default values as proposed by van Wagner (1987).}

\item{parallel}{Logical. Should parallel execution be used?}

\item{max.ncores}{Integer. Upper bound for user-defined number of cores.}

\item{ncores}{Integer number of cores used in parallel computation. Self-selected number of
 cores is used when \code{ncpus = NULL} (the default), or when \code{maxcores} exceeds the default \code{ncores} value.}
}
\value{
A vector of the same length as the input vectors (minus possible missing observations),
containing the requested components of the FWI system (either all or just FWI). See details.
}
\description{
Application of different bias correction methods to seasonal forecasts covering the World.
}
\section{Daylength adjustment factors}{

By default, the function applies the original FWI daylength adjustment factors for DC and DMC (van Wagner 1987),
 although it is possible to adjust them by as a function of latitude through the argument \code{lat}.
The reference values used for each latitudinal range are those indicated in p.71 and Tables A3.1 and A3.2 (p69) in
Lawson and Armitage (2008).
}
\author{
J. Bedia & M.Iturbide, partially based on the original FORTRAN code by van Wagner and Pickett (1985)
}
\references{
Lawson, B.D. & Armitage, O.B., 2008. Weather guide for the Canadian Forest Fire Danger Rating System. Northern Forestry Centre, Edmonton (Canada).

van Wagner, C.E., 1987. Development and structure of the Canadian Forest Fire Weather Index (Forestry Tech. Rep. No. 35). Canadian Forestry Service, Ottawa, Canada.

van Wagner, C.E., Pickett, T.L., 1985. Equations and FORTRAN program for the Canadian forest fire weather index system (Forestry Tech. Rep. No. 33). Canadian Forestry Service, Ottawa, Canada.
}

