% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fwiSkill.R
\name{fwiSkill}
\alias{fwiSkill}
\title{Skill calculation}
\usage{
fwiSkill(obs, fcst, verifun = c(""), prob, threshold = NULL,
  na.rm = FALSE, parallel = FALSE, ...)
}
\arguments{
\item{...}{same parameters as in function veriApply from package easyVerification}
}
\value{
A grid of the skill scores.
}
\description{
Skill calculation with function veriApply
}
\author{
M.Iturbide
}

